CI updates
- Use bionic distro to add support for PHP7.4 tests. Unfortunately,
trusty has some issues with PHP7.4 as it doesn't have zip and gd
enabled. I tried to manually add these via `pecl install` and it still
failed.
- Use nodejs 14
- Remove the uglify task since it was replaced by babel in 3.8+.
Unfortunately, the babel task does not work well with the plugin CI at
the moment. Dev should remember to manually verify this prior to
releasing new versions of the plugin.
- Unfortunately, mustache linting does not work with xenial distros and
above. I wanted to support PHP7.4 so I had to disable this at the
moment.